requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
black==23.9.1
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.103.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
mypy==1.5.1
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pydantic==2.3.0
pydantic_core==2.6.3
pytest==7.4.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
sqlalchemy-stubs==0.4
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
